Changed the index for invisible collision tiles. Added invisible collision tiles to negative space of all maps.
Quick explanation of how this works:
Collision 0 means empty space, anyone can walk or see through it.
Collision 1 (solid red) means an obstacle that can't be walked or seen through.
Collision 2 (solid blue) means an obstacle that can't be walked through but can be seen through (e.g. hole in the floor, short obstacle).
Collision 3 (dithered red) works like 1 but is not visible on the mini map
Collision 4 (dithered blue) works like 2 but is not visible on the mini map



git-svn-id: svn://localhost/media/c/Documnts/src/flare/repo.svndump/trunk@359 6ae5efe3-03d3-4ed7-8a64-3d71375a9a67